Avoid requestAnimationFrame
For reasons I don't fully understand, Chrome seems to stop running the
requestAnimationFrame callback after a while. This problem only seems to
manifest when Chrome is run in headless mode, and only when it's run by karma
(as opposed to fired up directly).

Anyway, rewrite to just use a setTimeout instead.